support ftp:// URLs in profiles
If the kernel or initramfs lines in the profile are of the form:

    ftp://<server>/<path>

then we use the z/VM FTP client to fetch the image from the given server
and path.

Any other string is assumed to be a TFTP path on the "PXE" server, so
existing profiles will continue to work the same.

This can be used to work around the 32MB file size limitation in the
z/VM TFTP client, which is a problem for booting Linux when the
initramfs image is larger than 32MB (for example, the Fedora 23
installer).